Update version to 3.5.1
Mirroring #1, I've updated the package versions to the latest versions available on conda-forge. 

As current version of `meta.yaml` prevents upgrade of `r-base` beyond version 3.4.1 unless the R version is manually specified in `conda install`, I suggest having the version dependencies for r-base and r-recommended be minimum instead of equals-to.